# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: IMAGINATE e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
  /published/{kind}:
    get:
      operationId: listPublishedContent
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Currently published documents
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ManagedDocument" } }
  /published/{kind}/{slug}:
    get:
      operationId: getPublishedContent
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Published document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedDocument" }
        "404": { description: Not published }
  /manage/{kind}:
    get:
      operationId: listManagedContent
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Admin documents
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ManagedDocument" } }
        "401": { description: Authentication required }
        "403": { description: Permission required }
    post:
      operationId: createManagedContent
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ManagedDocumentInput" }
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedDocument" }
        "400": { description: Invalid input }
        "403": { description: Permission required }
  /manage/{kind}/{id}:
    patch:
      operationId: updateManagedContent
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ManagedDocumentInput" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedDocument" }
        "400": { description: Invalid input }
        "403": { description: Permission required }
    delete:
      operationId: archiveManagedContent
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Archived }
...
components:
  schemas:
...
    ManagedDocumentInput:
      type: object
      required: [title, slug, status, data]
      properties:
        title: { type: string, minLength: 1, maxLength: 240 }
        slug: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        status: { type: string, enum: [draft, published, archived] }
        featured: { type: boolean }
        showInNavigation: { type: boolean }
        publishAt: { type: [string, "null"], format: date-time }
        unpublishAt: { type: [string, "null"], format: date-time }
        data: { type: object, additionalProperties: true }
    ManagedDocument:
      allOf:
        - { $ref: "#/components/schemas/ManagedDocumentInput" }
        - type: object
          required: [id, kind, createdAt, updatedAt]
          properties:
            id: { type: integer }
            kind: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...